requirements.txt
botcity-framework-core
pdfplumber
pypdf

